WIP: _split_into_chunks
