+++ environment.yml
@@ -1,0 +1,42 @@
+name: /mnt/P41/Repositories/extractor_SKKUGE/.conda
+channels:
+  - conda-forge
+  - bioconda
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=conda_forge
+  - _openmp_mutex=4.5=2_gnu
+  - black=22.10.0=py39hf3d152e_2
+  - bzip2=1.0.8=h7f98852_4
+  - ca-certificates=2022.12.7=ha878542_0
+  - click=8.1.3=unix_pyhd8ed1ab_2
+  - ld_impl_linux-64=2.39=hcc3a1bd_1
+  - libffi=3.4.2=h7f98852_5
+  - libgcc-ng=12.2.0=h65d4601_19
+  - libgomp=12.2.0=h65d4601_19
+  - libnsl=2.0.0=h7f98852_0
+  - libsqlite=3.40.0=h753d276_0
+  - libuuid=2.32.1=h7f98852_1000
+  - libzlib=1.2.13=h166bdaf_4
+  - mypy_extensions=0.4.3=py39hf3d152e_6
+  - ncurses=6.3=h27087fc_1
+  - openssl=3.0.7=h0b41bf4_1
+  - pathspec=0.10.3=pyhd8ed1ab_0
+  - pip=22.3.1=pyhd8ed1ab_0
+  - platformdirs=2.6.2=pyhd8ed1ab_0
+  - python=3.9.15=hba424b6_0_cpython
+  - python_abi=3.9=3_cp39
+  - readline=8.1.2=h0f457ee_0
+  - setuptools=65.6.3=pyhd8ed1ab_0
+  - tk=8.6.12=h27826a3_0
+  - tomli=2.0.1=pyhd8ed1ab_0
+  - typing-extensions=4.4.0=hd8ed1ab_0
+  - typing_extensions=4.4.0=pyha770c72_0
+  - tzdata=2022g=h191b570_0
+  - wheel=0.38.4=pyhd8ed1ab_0
+  - xz=5.2.6=h166bdaf_0
+  - pip:
+      - fastqsplitter==1.2.0
+      - isal==1.1.0
+      - xopen==1.7.0
+prefix: /mnt/P41/Repositories/extractor_SKKUGE/.conda
